requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
openai>=1.0.0
pydantic>=2.0.0
databricks-sdk>=0.20.0
